+++ packages/esm-app-shell/src/index.ts
@@ -1,16 +1,12 @@
 import "./layout";
 
-import {
-  ExtensionDefinition,
-  PageDefinition,
-  registerExtension,
-} from "@openmrs/esm-extension-manager";
-import { registerApplication, start } from "single-spa";
+import { start } from "single-spa";
+import { createAppState, setupApiModule } from "@openmrs/esm-api";
 import { setupI18n } from "./locale";
-import { routePrefix, routeRegex } from "./helpers";
+import { registerApp } from "./apps";
 import { sharedDependencies } from "./dependencies";
 import { loadModules, registerModules } from "./system";
-import type { SpaConfig, Activator, ActivatorDefinition } from "./types";
+import type { SpaConfig } from "./types";
 
 /**
  * Gets the microfrontend modules (apps). These are entries
@@ -33,27 +29,6 @@ function loadApps() {
 }
 
 /**
- * Normalizes the activator function, i.e., if we receive a
- * string we'll prepend the SPA base (prefix). We'll also handle
- * the case of a supplied array.
- * @param activator The activator to preprocess.
- */
-function preprocessActivator(
-  activator: ActivatorDefinition | Array<ActivatorDefinition>
-): Activator {
-  if (Array.isArray(activator)) {
-    const activators = activator.map(preprocessActivator);
-    return (location) => activators.some((activator) => activator(location));
-  } else if (typeof activator === "string") {
-    return (location) => routePrefix(activator, location);
-  } else if (activator instanceof RegExp) {
-    return (location) => routeRegex(activator, location);
-  } else {
-    return activator;
-  }
-}
-
-/**
  * Sets up the microfrontends (apps). Uses the defined export
  * from the root modules of the apps, which should export a
  * special function called "setupOpenMRS".
@@ -62,33 +37,7 @@ function preprocessActivator(
  */
 function setupApps(modules: Array<[string, System.Module]>) {
   for (const [appName, appExports] of modules) {
-    const setup = appExports.setupOpenMRS;
-
-    if (typeof setup === "function") {
-      const result = setup();
-
-      if (result && typeof result === "object") {
-        const availableExtensions: Array<ExtensionDefinition> =
-          result.extensions ?? [];
-
-        const availablePages: Array<PageDefinition> = result.pages ?? [];
-
-        if (typeof result.activate !== "undefined") {
-          availablePages.push({
-            load: result.lifecycle,
-            route: result.activate,
-          });
-        }
-
-        for (const { name, load } of availableExtensions) {
-          registerExtension(appName, name, load);
-        }
-
-        for (const { route, load } of availablePages) {
-          registerApplication(appName, load, preprocessActivator(route));
-        }
-      }
-    }
+    registerApp(appName, appExports);
   }
 }
 
@@ -132,7 +81,7 @@ function renderDevResetButton() {
 }
 
 function clearDevOverrides() {
-  for (let key of Object.keys(localStorage)) {
+  for (const key of Object.keys(localStorage)) {
     if (
       key.startsWith("import-map-override:") &&
       !["import-map-override:react", "import-map-override:react-dom"].includes(
@@ -152,5 +101,7 @@ function clearDevOverrides() {
 export function initializeSpa(config: SpaConfig) {
   setupPaths(config);
   registerModules(sharedDependencies);
+  setupApiModule();
+  createAppState({});
   return loadApps().then(setupApps).then(runShell).catch(handleInitFailure);
 }
